+++ V1/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class V1Config(AppConfig):
+    name = 'V1'
