+++ V1/devices/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class DevicesConfig(AppConfig):
+    name = 'devices'
