+++ V1/devices/serializers.py
@@ -1,0 +1,10 @@
+from rest_framework import serializers
+from V1.devices.models import Device
+
+class DeviceSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Device
+        fields = [
+            'id',
+            'sms_number',
+        ]
