+++ V1/locations/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class LocationsConfig(AppConfig):
+    name = 'locations'
