+++ V1/locations/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path, include
+from rest_framework.urlpatterns import format_suffix_patterns
+# from .views import DeviceViews, UserViews
+
+urlpatterns = [
+
+]
