+++ machine/settings.py
@@ -45,7 +45,7 @@ INSTALLED_APPS = [
     'rest_framework',
     'corsheaders',
     'dj_database_url',
-    'api',
+    'V1',
 ]
 
 MIDDLEWARE = [
